Actually fix crash
